+++ packages/epub_view/example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,8 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import wakelock_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  WakelockMacosPlugin.register(with: registry.registrar(forPlugin: "WakelockMacosPlugin"))
 }