+++ packages/epub_view/lib/epub_view.dart
@@ -1,4 +1,6 @@
 export 'package:epubx/epubx.dart';
-export 'src/epub_view.dart';
+
 export 'src/ui/actual_chapter.dart';
+export 'src/ui/epub_view.dart';
 export 'src/ui/table_of_contents.dart';
+export 'src/helpers/epub_document.dart';