+++ packages/epub_view/lib/src/data/models/chapter.dart
@@ -1,0 +1,15 @@
+class EpubViewChapter {
+  EpubViewChapter(this.title, this.startIndex);
+
+  final String? title;
+  final int startIndex;
+
+  String get type => this is EpubViewSubChapter ? 'subchapter' : 'chapter';
+
+  @override
+  String toString() => '$type: {title: $title, startIndex: $startIndex}';
+}
+
+class EpubViewSubChapter extends EpubViewChapter {
+  EpubViewSubChapter(String? title, int startIndex) : super(title, startIndex);
+}