+++ packages/epub_view/lib/src/data/models/paragraph.dart
@@ -1,0 +1,8 @@
+import 'package:html/dom.dart' as dom;
+
+class Paragraph {
+  Paragraph(this.element, this.chapterIndex);
+
+  final dom.Element element;
+  final int chapterIndex;
+}