+++ packages/epub_view/lib/src/data/models/parse_result.dart
@@ -1,0 +1,11 @@
+import 'package:epub_view/src/data/epub_parser.dart';
+
+export 'package:epubx/epubx.dart' hide Image;
+
+class ParseResult {
+  const ParseResult(this.epubBook, this.chapters, this.parseResult);
+
+  final EpubBook epubBook;
+  final List<EpubChapter> chapters;
+  final ParseParagraphsResult parseResult;
+}